Change $cpt_supports value from array() to false
In Advanced Options: Supports, when nothing is checked, the default title and editor will still show up. Set this to false to fix the issue.

From register_post_type reference: As of 3.5, boolean false can be passed as value instead of an array to prevent default (title and editor) behavior.