Rewrite changelog.txt as CHANGELOG.md
Uses the de facto standard from keepachangelog.com. `CHANGELOG.md` is also more of a standard than `changelog.txt`, since the uppercase means it can jump to the top of the files in a case-sensitive operating system, and the markdown allows for richer organisation of the lists of changes.

- Headings for each release, allowing easier jumping to specific sections.
- Subsections for each change type.
- Each version number linked to a comparison of commits.
- Dates consistently formatted as `YYYY-MM-DD`.
- Technical terms / code marked with backticks.
- Typos fixed.

See #658.