Docs: Remove change log, point to CHANGELOG.md
Anyone reading this `README.md` (rather than `readme.txt`) is likely reading the source files, or the project on GitHub. In which case, keeping the list of changes in one place, and referring to that canonical location is more effective in terms of atomic commits and intuition about where to look for changes.

See #658.